requirements.txt
langchain-ollama
langchain
chromadb
streamlit
langchain-community
langchain-chroma
pymupdf
langchain-experimental
google-search-results